requirements.txt
telethon
elasticsearch
pandas
sklearn
sentiment_analysis_spanish
pyyaml